requirement.txt
allure-behave==2.6.2
allure-pytest==2.5.4
allure-python-commons==2.6.2
apipkg==1.5
Appium-Python-Client==0.36
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.6
behave==1.2.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
chromedriver==2.24.1
Click==7.0
execnet==1.5.0
Flask==1.0.2
gevent==1.3.7
greenlet==0.4.15
html-testRunner==1.1.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.9.5
MarkupSafe==1.1.0
more-itertools==5.0.0
msgpack==0.6.0
nose==1.3.7
parse==1.12.0
parse-type==0.4.2
pluggy==0.8.1
progress-bar==8
progressbar==2.5
public==2019.4.13
py==1.7.0
pycparser==2.19
pytest==4.1.1
pytest-forked==1.0.1
pytest-parallel==0.0.9
pytest-repeat==0.8.0
pytest-xdist==1.26.0
python-dotenv==0.10.1
pyzmq==17.1.2
requests==2.21.0
roaster==1.0
robotframework==3.1.1
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.3.1
selenium==3.141.0
six==1.12.0
splinter==0.10.0
timedelta==2019.4.13
unittest-xml-reporting==2.2.1
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
